requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
slowapi>=0.1.5
pydantic>=1.8.2
jinja2>=3.0.1
aiofiles>=0.7.0
speedtest-cli>=2.1.3
